refactor(jazz-next): extract module replay from driver
